+++ streamingpro-spark-2.4.0-adaptor/src/main/java/org/apache/spark/MLSQLSyntaxErrorListener.java
@@ -1,0 +1,13 @@
+package org.apache.spark;
+
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
+
+public class MLSQLSyntaxErrorListener extends BaseErrorListener {
+
+    @Override
+    public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line, int charPositionInLine, String msg, RecognitionException e) {
+        throw new RuntimeException("MLSQL Parser error in [row:" + line + " column:" + charPositionInLine + "]  error msg: " + msg);
+    }
+}
